fix(core,quant,eval): round-180 deep-dive correctness fixes (TASK-341/342/343, ISS-383..386)
- RMSNorm (ISS-384): mean-square was accumulated in host dtype — fp16
  pow(2) overflows past |x|~256 and the layer silently collapsed to
  all-zeros (verified). Accumulate in fp32, cast back for the RMS.
- Fp8QuantizedLinear (ISS-383): output dtype resolved from the fp32 bias,
  so biased FP8 layers emitted fp32 into an fp16 graph (verified). Now
  resolves from the input (the graph dtype), matching GPTQ/AWQ/Smooth.
- PerplexityMetric (ISS-385/386): registry resolved ignore_index=None
  (metrics/__init__ documents -100) so an all-ignored corpus returned NaN
  not inf; and references narrower than predictions crashed with a raw
  batch-mismatch ValueError. Default is now -100 and both sides are
  clipped to a common label_width window.

All RED-checked before fix; full affected suites green.

+++ src/llm/core/rms_norm.py
@@ -73,7 +73,13 @@ class RMSNorm(nn.Module):
 
         # 计算均方值 (不减去均值的方差)
         # mean(x^2)
-        mean_square = torch.mean(hidden_states.pow(2), dim=dims_to_normalize, keepdim=True)
+        # RIL ISS-384: ``x.pow(2)`` in host dtype overflows past |x| > ~256 for
+        # fp16 (max 65504) and ~3.4e38/bf16 is even lower in practice, turning
+        # the mean-square inf and collapsing the layer to all-zeros. Accumulate
+        # in fp32 (standard RMSNorm practice) and keep the RMS in fp32: the
+        # subsequent ``hidden_states / rms`` division upcasts the numerator,
+        # so fp32 RMS is both correct and free.
+        mean_square = torch.mean(hidden_states.float().pow(2), dim=dims_to_normalize, keepdim=True)
 
         # 计算 RMS: sqrt(mean(x^2) + eps)
         rms = torch.sqrt(mean_square + self.eps)
